refactor(dismissible-box): convert files to typescript
