--- src/components/dismissible-box/index.d.ts
@@ -1,1 +1,0 @@
-export { default } from "./dismissible-box";
