--- src/components/dismissible-box/index.js
@@ -1,1 +1,0 @@
-export { default } from "./dismissible-box.component";
