+++ src/components/dismissible-box/index.ts
@@ -1,0 +1,2 @@
+export type { DismissibleBoxProps } from "./dismissible-box.component";
+export { default } from "./dismissible-box.component";
